revert(storybook): contexts addon, since it's missing the core-js dependency
