+++ packages/application-components/.storybook/addons.js
@@ -5,4 +5,3 @@ import '@storybook/addon-storysource/register';
 import '@storybook/addon-actions/register';
 import '@storybook/addon-options/register';
 import '@storybook/addon-links/register';
-import '@storybook/addon-contexts/register';
