--- packages/application-components/.storybook/configs/contexts/index.js
@@ -1,3 +1,0 @@
-import intlContext from './intl-context';
-
-export const contexts = [intlContext];
