+++ packages/application-components/.storybook/decorators/index.js
@@ -1,1 +1,2 @@
+export { default as Intl } from './section';
 export { default as Section } from './section';
