+++ packages/application-components/.storybook/decorators/intl/index.js
@@ -1,0 +1,1 @@
+export { default } from './intl';
