+++ packages/application-components/package.json
@@ -50,7 +50,6 @@
   "devDependencies": {
     "@commercetools-frontend/ui-kit": "9.8.3",
     "@storybook/addon-actions": "5.1.3",
-    "@storybook/addon-contexts": "5.1.3",
     "@storybook/addon-info": "5.1.3",
     "@storybook/addon-knobs": "5.1.3",
     "@storybook/addon-links": "5.1.3",
